Revert "PYIC-7872: Add AppConfigService"
