+++ gradle/libs.versions.toml
@@ -16,7 +16,6 @@ awsSdkDynamodbEnhanced = { module = "software.amazon.awssdk:dynamodb-enhanced" }
 awsSdkKms = { module = "software.amazon.awssdk:kms" }
 awsSdkSqs = { module = "software.amazon.awssdk:sqs" }
 awsSdkUrlConnectionClient = { module = "software.amazon.awssdk:url-connection-client" }
-awsSdkAppConfigData = { module = "software.amazon.awssdk:appconfigdata" }
 commonsCodec = "commons-codec:commons-codec:1.18.0"
 commonsCollections = "org.apache.commons:commons-collections4:4.5.0-M2"
 diVocab = "uk.gov.di.model:di-data-model-jackson:1.9.1"
