+++ libs/common-services/build.gradle
@@ -15,7 +15,6 @@ dependencies {
 			libs.awsSdkDynamodb,
 			libs.awsSdkDynamodbEnhanced,
 			libs.awsSdkKms,
-			libs.awsSdkAppConfigData,
 			libs.commonsCodec,
 			libs.commonsCollections,
 			libs.jacksonDatabind,
