+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/config/EnvironmentVariable.java
@@ -1,13 +1,9 @@
 package uk.gov.di.ipv.core.library.config;
 
 public enum EnvironmentVariable {
-    APP_CONFIG_ENVIRONMENT_ID,
-    APP_CONFIG_ID,
-    APP_CONFIG_PROFILE_ID,
     CLIENT_AUTH_JWT_IDS_TABLE_NAME,
     CLIENT_OAUTH_SESSIONS_TABLE_NAME,
     CONFIG_SERVICE_CACHE_DURATION_MINUTES,
-    CONFIG_SOURCE,
     CRI_OAUTH_SESSIONS_TABLE_NAME,
     CRI_RESPONSE_TABLE_NAME,
     ENVIRONMENT,
