+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/service/SsmConfigService.java
@@ -57,8 +57,10 @@ public class SsmConfigService extends ConfigService {
     @ExcludeFromGeneratedCoverageReport
     public SsmConfigService() {
         var cacheDuration =
-                getIntegerEnvironmentVariable(
-                        CONFIG_SERVICE_CACHE_DURATION_MINUTES, DEFAULT_CACHE_DURATION_MINUTES);
+                getEnvironmentVariable(CONFIG_SERVICE_CACHE_DURATION_MINUTES) == null
+                        ? DEFAULT_CACHE_DURATION_MINUTES
+                        : Integer.parseInt(
+                                getEnvironmentVariable(CONFIG_SERVICE_CACHE_DURATION_MINUTES));
 
         this.ssmProvider =
                 ParamManager.getSsmProvider(
